+++ src/Components/App/App.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { render, screen } from '@testing-library/react'
-import App from 'components/App'
+import App from 'components/app'
 
 test('renders learn react link', () => {
   render(<App />)
