+++ src/Components/Main/index.scss
