+++ src/Components/Main/index.tsx
@@ -1,12 +1,14 @@
 import { Container } from '@mui/material'
-import { LocationForm } from 'components/LocationForm'
-import 'components/Main/Main.scss'
+import { LocationForm } from 'components/locationForm'
+import 'components/main/index.scss'
+import { SatelliteTable } from 'components/satellitesTable'
 
 export const Main = () => {
   return (
     <Container fixed>
         <h1>Input your location</h1>
         <LocationForm />
+        <SatelliteTable />
     </Container>
   )
 }
