+++ src/index.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import ReactDOM from 'react-dom'
 import 'index.css'
 import reportWebVitals from 'reportWebVitals'
-import App from 'components/App'
+import App from 'components/app'
 
 ReactDOM.render(
   <React.StrictMode>
